error benchmark

+++ cpp/src/main.cpp
@@ -6,7 +6,7 @@
 
 PYBIND11_MODULE(parinterp, m) {
     py::class_<Interpolator>(m, "Linear2DInterpolatorCpp", "Interpolator class")
-        .def(py::init<const Triangulator::pyarr_size_t&, size_t>())
+        .def(py::init<const Triangulator::pyarr_size_t&, int>())
         .def("__call__", &Interpolator::operator());
 }
